Setup experiment for IK. Clean up interfaces
