+++ Cargo.toml
@@ -15,13 +15,13 @@ description = "Direkt and inverse kinematics for biomechanics, robotics and char
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-faer = { version = "0.19.4", optional = true }
-itertools = "0.13.0"
+faer = { version = "0.21.7", optional = true }
+itertools = "0.14.0"
 nalgebra = { version = "0.33.2", optional = true }
 ndarray = { version = "0.16.1", optional = true, features = ["approx", "rayon"] }
 num-traits = "0.2.19"
 rayon = { version = "1.10.0", optional = true }
-thiserror = "2.0.7"
+thiserror = "2.0.12"
 tracing = "0.1.41"
 tracing-attributes = "0.1.28"
 
@@ -33,4 +33,5 @@ rayon = ["dep:rayon"]
 
 [dev-dependencies]
 approx = "0.5"
-test-log = { version = "0.2.16", features = ["trace"] }
+mannequin = { path = ".", default-features = false, features = ["ndarray"] }
+test-log = { version = "0.2.17", features = ["trace"] }
