+++ src/ndarray/robot.rs
@@ -205,107 +205,4 @@ pub struct DifferentialIK {
 // pub type BasicMannequin = Mannequin<DirectedArenaTree<Bone, LinkNodeId>, Bone, ForwardsKinematics, DifferentialIK>;
 
 #[cfg(test)]
-mod tests {
-    use super::{Axis, DifferentialIK, LinkNodeId, Segment};
-    use crate::differentiable::ComputeSelection;
-    use crate::{DepthFirstArenaTree, DifferentiableModel, DirectedArenaTree, DirectionIterable, Forward, Rigid};
-    use crate::{Differentiable, ForwardModel};
-    use approx::assert_abs_diff_eq;
-    use itertools::Itertools;
-    use ndarray::{prelude::*, Order};
-
-    #[test]
-    fn test_fk() {
-        let mut tree = DirectedArenaTree::new();
-        let mut fk = ForwardModel::new(10, DifferentiableModel::new());
-
-        let mut trafo = Segment::neutral_element();
-        trafo.slice_mut(s![..3, 3]).assign(&array![10.0, 0.0, 0.0]);
-
-        let link1 = Segment::new(&trafo, Axis::RotationZ, None);
-        let link2 = Segment::new(&trafo, Axis::RotationZ, Some(Segment::neutral_element()));
-        let link3 = Segment::new(&trafo, Axis::RotationZ, Some(Segment::neutral_element()));
-        let link4 = Segment::new(&trafo, Axis::RotationZ, Some(Segment::neutral_element()));
-
-        // TODO .. can we make the refs fix in a way they don't get optimized away?
-        // Then these could be strings even!
-        let ref1 = tree.set_root(link1, "link1".to_string());
-        let ref2 = tree.add(link2, "link2".to_string(), &ref1).unwrap();
-        let ref3 = tree.add(link3, "link3".to_string(), &ref1).unwrap();
-        let ref4 = tree.add(link4, "link4".to_string(), &ref3).unwrap();
-
-        let tree: DepthFirstArenaTree<_, _> = tree.into();
-
-        fk.initialize(&tree, &[&ref1, &ref2, &ref3, &ref4], &[&ref2, &ref3, &ref4]);
-        let res = fk.solve(&tree, &[0.0, 0.0, std::f64::consts::FRAC_PI_2, 0.0]);
-        let res = res.iter().map(|&el| el.to_owned()).collect_vec();
-
-        assert_eq!(
-            res,
-            vec![vec![20.0, 0.0, 0.0], vec![20.0, 0.0, 0.0], vec![20.0, 10.0, 0.0]]
-        );
-    }
-
-    #[test]
-    fn test_jacobian() {
-        let ik = DifferentialIK { max_depth: 10 };
-        let mut tree = DirectedArenaTree::<Segment, LinkNodeId>::new();
-
-        let mut trafo = Segment::neutral_element();
-        trafo.slice_mut(s![..3, 3]).assign(&array![10.0, 0.0, 0.0]);
-
-        let link1 = Segment::new(&trafo, Axis::RotationZ, None);
-        let link2 = Segment::new(&trafo, Axis::RotationZ, Some(trafo.clone()));
-        let link3 = Segment::new(&trafo, Axis::RotationZ, None);
-        let link4 = Segment::new(&trafo, Axis::RotationZ, Some(trafo.clone()));
-        // Doesn't do anything, is at the end
-        let link5 = Segment::new(&trafo, Axis::RotationZ, Some(trafo.clone()));
-
-        // TODO .. can we make the refs fix in a way they don't get optimized away?
-        // Then these could be strings even!
-
-        let ref1 = tree.set_root(link1, "link1".to_string());
-        let ref2 = tree.add(link2, "link2".to_string(), &ref1).unwrap();
-        let ref3 = tree.add(link3, "link3".to_string(), &ref1).unwrap();
-        let ref4 = tree.add(link4, "link4".to_string(), &ref3).unwrap();
-        tree.add(link5, "link5".to_string(), &ref4).unwrap();
-        let tree: DepthFirstArenaTree<_, _> = tree.into();
-
-        let mut jacobian = DifferentiableModel::<f64>::new();
-
-        jacobian.setup(
-            &tree,
-            &[
-                &"link1".to_string(),
-                &"link2".to_string(),
-                &"link3".to_string(),
-                &"link4".to_string(),
-            ]
-            .into(),
-            // TODO automatically those that have an effector defined
-            &[&"link2".to_string(), &"link4".to_string()].into(),
-        );
-
-        jacobian.compute(
-            &tree,
-            &[0.0, 0.0, std::f64::consts::FRAC_PI_2, std::f64::consts::FRAC_PI_2, 0.0],
-            ComputeSelection::JacobianOnly,
-        );
-
-        let result = ArrayView1::<f64>::from(jacobian.jacobian())
-            .into_shape_with_order(((jacobian.rows(), jacobian.cols()), Order::ColumnMajor))
-            .unwrap();
-
-        let target = array![
-            [0.0, 0.0, 0.0, 0.0],
-            [20.0, 10.0, 0.0, 0.0],
-            [0.0, 0.0, 0.0, 0.0,],
-            [-10.0, 0.0, -10.0, 0.0],
-            [0.0, 0.0, -10.0, -10.0],
-            [0.0, 0.0, 0.0, 0.0,]
-        ];
-
-        assert_eq!(jacobian.dims(), (6, 4));
-        assert_abs_diff_eq!(result, target, epsilon = 1e-6);
-    }
-}
+mod tests {}
